@@ -12,25 +12,23 @@
     "description": "A coroutine framework that focuses on hyperspeed and flexible, specifically use for build microservices and middlewares.",
     "license": "Apache-2.0",
     "require": {
-        "php": ">=7.3",
-        "ext-swoole": ">=4.7",
+        "php": ">=8.1",
+        "ext-swoole": ">=5.0",
         "ext-json": "*",
-        "hyperf/framework": "~2.2.0",
-        "hyperf/crontab": "~2.2.0",
-        "hyperf/task": "~2.2.0",
-        "hyperf/websocket-server": "~2.2.0",
-        "hyperf/rpc-server": "~2.2.0",
-        "dleno/common-core": "~2.2.0"
+        "hyperf/framework": "~3.1.0",
+        "hyperf/crontab": "~3.1.0",
+        "hyperf/websocket-server": "~3.1.0",
+        "dleno/common-core": "~3.1.0"
     },
     "require-dev": {
-        "swoole/ide-helper": "^4.6",
         "friendsofphp/php-cs-fixer": "^3.0",
+        "hyperf/devtool": "~3.1.0",
+        "hyperf/testing": "~3.1.0",
         "mockery/mockery": "^1.0",
-        "phpstan/phpstan": "^0.12",
-        "hyperf/devtool": "~2.2.0",
-        "hyperf/testing": "~2.2.0",
+        "phpstan/phpstan": "^1.0",
+        "swoole/ide-helper": "^5.0",
         "squizlabs/php_codesniffer": "^3.5",
-        "hyperf/ide-helper": "~2.2.0"
+        "hyperf/ide-helper": "~3.1.0"
     },
     "suggest": {
         "ext-openssl": "Required to use HTTPS.",
@@ -44,7 +42,8 @@
     "autoload": {
         "psr-4": {
             "App\\": "app/"
-        }
+        },
+        "files": []
     },
     "autoload-dev": {
         "psr-4": {
@@ -53,6 +52,11 @@
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
+    "config": {
+        "secure-http": false,
+        "optimize-autoloader": true,
+        "sort-packages": true
+    },
     "extra": [],
     "scripts": {
         "post-root-package-install": [
@@ -68,10 +72,5 @@
             "Composer\\Config::disableProcessTimeout",
             "php ./bin/hyperf.php start"
         ]
-    },
-    "config": {
-        "secure-http": false,
-        "optimize-autoloader": true,
-        "sort-packages": true
     }
 }
