@@ -7,20 +7,12 @@ namespace App\Amqp\Producer;
 use Hyperf\Amqp\Annotation\Producer;
 use Dleno\CommonCore\Base\Amqp\BaseProducer;
 
-/**
- * @Producer(exchange="", routingKey="")
- */
+#[Producer(exchange:"", routingKey:"")]
 class TestProducer extends BaseProducer
 {
-    /**
-     * @var string
-     */
-    protected $exchange = 'TestExchange';//交换机key(注解优先，需要动态设置时，则不能要注解)
+    protected string$exchange = 'TestExchange';//交换机key(注解优先，需要动态设置时，则不能要注解)
 
-    /**
-     * @var string
-     */
-    protected $routingKey = 'TestRouting';//路由key(注解优先，需要动态设置时，则不能要注解)
+    protected array|string $routingKey = 'TestRouting';//路由key(注解优先，需要动态设置时，则不能要注解)
 
     /**
      * 是否延迟消息交换机(生产者消费者要对应)
