@@ -2,8 +2,6 @@
 
 namespace App\Aspect;
 
-use App\Components\Admin\BlindBox\BlindBoxComponent;
-use App\Conf\ManagerRedisKeyConf;
 use App\Conf\SignConf;
 use App\Tools\ApiServer;
 use Hyperf\Di\Annotation\Aspect;
@@ -12,7 +10,7 @@ use Hyperf\Di\Aop\AbstractAspect;
 use Hyperf\Di\Aop\ProceedingJoinPoint;
 use Hyperf\HttpServer\Contract\RequestInterface;
 use Hyperf\Redis\Redis;
-use Hyperf\Utils\Context;
+use Hyperf\Context\Context;
 use Dleno\CommonCore\Conf\GlobalConf;
 use Dleno\CommonCore\Conf\RcodeConf;
 use Dleno\CommonCore\Exception\Http\HttpException;
@@ -22,25 +20,22 @@ use Dleno\CommonCore\Tools\Logger;
 use Dleno\CommonCore\Tools\Server;
 use Psr\Http\Message\ServerRequestInterface;
 
-/**
- * @Aspect
- */
+use function Hyperf\Config\config;
+
+#[Aspect]
 class AdminModuleBeforeAspect extends AbstractAspect
 {
     // 要切入的类，可以多个，亦可通过 :: 标识到具体的某个方法，通过 * 可以模糊匹配
-    public $classes = [];
+    public array $classes = [];
 
     // 要切入的注解，具体切入的还是使用了这些注解的类，仅可切入类注解和类方法注解
-    public $annotations = [
+    public array $annotations = [
         \Hyperf\HttpServer\Annotation\AutoController::class,
         \Hyperf\HttpServer\Annotation\Controller::class,
     ];
 
-    /**
-     * @Inject()
-     * @var Redis
-     */
-    public $redis;
+    #[Inject]
+    public Redis $redis;
 
     public function process(ProceedingJoinPoint $proceedingJoinPoint)
     {
@@ -169,7 +164,14 @@ class AdminModuleBeforeAspect extends AbstractAspect
      */
     private function checkAuth($whiteVal)
     {
-        $token = get_header_val('Client-Token', 0);;
-//        get_inject_obj(BlindBoxComponent::class)->checkAuth($token);
+        //白名单
+        if (CheckVal::checkInStatus(GlobalConf::WHITE_TYPE_TOKEN, $whiteVal)) {
+            return;
+        }
+        /*$token = get_header_val('Client-Token', 0);
+        $checkAuth = get_inject_obj(BlindBoxComponent::class)->checkAuth($token);
+        if (!$checkAuth) {
+            throw new HttpException('Error Sign', RcodeConf::ERROR_TOKEN);
+        }*/
     }
 }
