@@ -16,17 +16,17 @@ use Dleno\CommonCore\Tools\Output\ApiOutLog;
 use Dleno\CommonCore\Tools\Server;
 use Psr\Http\Message\ResponseInterface;
 
-/**
- * @Aspect
- */
+use function Hyperf\Config\config;
+
+#[Aspect]
 class ApiOutputAspect extends AbstractAspect
 {
     // 要切入的类，可以多个，亦可通过 :: 标识到具体的某个方法，通过 * 可以模糊匹配
-    public $classes = [
+    public array $classes = [
     ];
 
     // 要切入的注解，具体切入的还是使用了这些注解的类，仅可切入类注解和类方法注解
-    public $annotations = [
+    public array $annotations = [
         \Hyperf\HttpServer\Annotation\AutoController::class,
         \Hyperf\HttpServer\Annotation\Controller::class,
         \Dleno\CommonCore\Annotation\ExceptionHandlerLog::class,
