@@ -7,15 +7,17 @@ namespace App\AsyncQueue\Process;
 use Dleno\CommonCore\Base\AsyncQueue\BaseQueueConsumer;
 use Hyperf\Process\Annotation\Process;
 
+use function Hyperf\Config\config;
+
 /**
  * 默认消费进程（使用默认队列的都在此消费）
- * @Process(nums=2)
  */
+#[Process(nums:2)]
 class DefaultQueueConsumer extends BaseQueueConsumer
 {
-    protected $queue = 'default';
+    protected string $queue = 'default';
 
-    protected $reloadChannel = ['timeout', 'failed'];
+    protected array $reloadChannel = ['timeout', 'failed'];
 
     public function isEnable($server): bool
     {
