@@ -7,17 +7,15 @@ use Hyperf\Di\Aop\AbstractAspect;
 use Hyperf\Di\Aop\ProceedingJoinPoint;
 use Dleno\CommonCore\Tools\Output\WsOutLog;
 
-/**
- * @Aspect
- */
+#[Aspect]
 class ApiOutputAspect extends AbstractAspect
 {
     // 要切入的类，可以多个，亦可通过 :: 标识到具体的某个方法，通过 * 可以模糊匹配
-    public $classes = [
+    public array $classes = [
     ];
 
     // 要切入的注解，具体切入的还是使用了这些注解的类，仅可切入类注解和类方法注解
-    public $annotations = [
+    public array $annotations = [
         \Dleno\CommonCore\Annotation\WsController::class,
         \Dleno\CommonCore\Annotation\WsExceptionHandlerLog::class,
     ];
