@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace App\Components;
 
 use Dleno\CommonCore\Base\BaseCoreComponent;
+use Hyperf\Di\Annotation\Inject;
+use Hyperf\Redis\Redis;
 
 /**
  * Class BaseComponent
@@ -12,5 +14,6 @@ use Dleno\CommonCore\Base\BaseCoreComponent;
  */
 class BaseComponent extends BaseCoreComponent
 {
-
+    #[Inject]
+    protected Redis $redis;
 }
