@@ -5,10 +5,13 @@ declare(strict_types=1);
  * This file is part of Hyperf.
  *
  * @link     https://www.hyperf.io
- * @document https://doc.hyperf.io
+ * @document https://hyperf.wiki
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
+use function Hyperf\Support\env;
+use function Hyperf\Support\value;
+
 return [
     'default' => value(
         function ($poolName = 'default') {
@@ -26,7 +29,6 @@ return [
                 'params'     => $params,
                 'open_ssl'   => env('AMQP_OPEN_SSL', false),
             ];
-
             return $config;
         }
     ),
@@ -49,7 +51,6 @@ return [
                 'params'     => $params,
                 'open_ssl'   => env('AMQP_OPEN_SSL', false),
             ];
-
             return $config;
         }
     ),
