@@ -1,8 +1,17 @@
 <?php
 
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://hyperf.wiki
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
 return [
     'default' => [
-        'driver'         => Hyperf\AsyncQueue\Driver\RedisDriver::class,
+        'driver'         => \Hyperf\AsyncQueue\Driver\RedisDriver::class,
         'redis'          => [
             'pool' => 'default'
         ],
