@@ -1,10 +1,14 @@
 <?php
 
 declare(strict_types=1);
-
-use Monolog\Formatter;
-use Monolog\Logger;
-
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://hyperf.wiki
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
 return array_merge(
     \Dleno\CommonCore\Logger\LoggerConfig::getDefaultConfig(),
     [
