@@ -9,6 +9,8 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
+use function Hyperf\Support\env;
+
 $exceptions = [];
 if (env('ENABLE_HTTP', false)) {
     $exceptions['http'] = [
@@ -18,11 +20,6 @@ if (env('ENABLE_HTTP', false)) {
 if (env('ENABLE_WS', false)) {
     $exceptions['ws'] = [
         \App\WebSocket\Middleware\WebSocketAuthMiddleware::class,//ws握手
-    ];
-}
-if (env('ENABLE_RPC', false)) {
-    $exceptions['jsonrpc'] = [
-        \Dleno\CommonCore\Middleware\Rpc\InitMiddleware::class,//初始化中间件
     ];
 }
 return $exceptions;
