@@ -12,13 +12,10 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
+use Hyperf\Context\ApplicationContext;
 use Hyperf\Di\Container;
 use Hyperf\Di\Definition\DefinitionSourceFactory;
-use Hyperf\Utils\ApplicationContext;
 
-$container = new Container((new DefinitionSourceFactory(true))());
+$container = new Container((new DefinitionSourceFactory())());
 
-if (! $container instanceof \Psr\Container\ContainerInterface) {
-    throw new RuntimeException('The dependency injection container is invalid.');
-}
 return ApplicationContext::setContainer($container);
